+++ vm/src/vm/runners/cairo_runner.rs
@@ -48,7 +48,7 @@ use crate::{
 };
 use num_integer::div_rem;
 use num_traits::{ToPrimitive, Zero};
-use serde::Deserialize;
+use serde::{Deserialize, Serialize};
 
 use super::{
     builtin_runner::{KeccakBuiltinRunner, PoseidonBuiltinRunner, OUTPUT_BUILTIN_NAME},
@@ -1292,7 +1292,7 @@ pub struct SegmentInfo {
 //*   ExecutionResources
 //* ----------------------
 
-#[derive(Clone, Debug, Default, Deserialize, Eq, PartialEq)]
+#[derive(Clone, Debug, Default, Deserialize, Serialize, Eq, PartialEq)]
 pub struct ExecutionResources {
     pub n_steps: usize,
     pub n_memory_holes: usize,